cd into newly cloned git repositories
`cd` doesn't work in shell scripts because each one runs in its own subshell.
The `cd` successfully changes the directory *in the subshell*, then the subshell
dies because the shell script has finished, and nothing happens in your shell.

So `superclone` now returns the name of the directory to `cd` into and we run
`cd` as a function in *this* shell.

It's all very silly.