We only need to grab the git status once
Instead of grabbing the git status, using it to generate a letter like "P", then
passing the "P" to another function that colorizes it based on the git status,
we instead grab the git status once and directly generate a colored "P".

Remove "pending" status since I don't know how to trigger it and it doesn't get
colorized anyway.