Use rcm's tag feature to split out Haskell config
Splitting out the Zsh and Vim configuration required some acrobatics.

Previously, we instructed rcm not to recurse into ~/.zsh and ~/.vim, just
symlink ~/.zsh to zsh and ~/.vim to vim (with the SYMLINK_DIRS option in rcrc).

Now that there are files in tag-haskell that need to be deeply nested under zsh/
and vim/, we tell rcm to recurse into them and not just treat them as solid
chunks by removing them from SYMLINK_DIRS. If they were still symlinked, then:

* rcm would link tag-haskell/zsh/haskell.zsh -> ~/.zsh/haskell.zsh
* But ~/.zsh is a symlink
* So really it would be tag-haskell/zsh/haskell.zsh -> ~/.dotfiles/zsh/haskell.zsh
* And boom, suddenly there's a new, unwanted symlink in my git repo

To handle vim-plug, we:

* start vim-plug
* source plugins.vim for general plugins
* source ~/.vim/plugins/*.vim so that any tags can put their own vim plugins in
  tag-whatever/vim/plugins/whatever.vim
* finish vim-plug