Split out tmux configuration
The very first thing the zshrc should do is connect to tmux.

If it keeps loading things in the non-tmux "background" after connecting to
tmux, I don't care, I'm in tmux. But I don't want it to load 18 things _before_
connecting me to tmux.

In order to allow running things early in the zsh boot process, introduce
`zsh/before/` and `zsh/after/` directories, which currently only tmux uses. It
uses `zsh/before/` to boot tmux early in the process, and `zsh/after/` to go to
the tmux session's current directory after booting.