Smartly search for the word under the cursor
Sometimes the word under the cursor includes punctuation, in which case
'\bWORD!\b' will fail because \b is a word boundary and there's a non-word
character in WORD. So, remove them. This results in a less-precise
match (it'll find WORD as well as WORD!, for example), but is better
than getting zero results.

The non-word characters are always at the beginning (`$variable`) or end
(`function!`) of a word, so it's fine to remove them and search for the
suffix or prefix.