Use OSX's built-in Zsh
I recently restored my files to a new computer, but didn't copy over
/usr/local.

When I signed in, my $SHELL was set to /usr/local/bin/zsh, which didn't
exist, and so Terminal.app exited immediately upon opening. I had to
use iTerm (with a custom profile that overrode my user's $SHELL) in
order to log in and change my shell back to /bin/zsh in order to do
anything.

Obviously this was a total nightmare. To fix it:

* Don't install Zsh via Homebrew at all. OSX ships with Zsh 5.2 now,
  which a comment from Past Gabe implies was the whole reason to use
  Homebrew-provided Zsh.
* Set my shell to the ever-present /bin/zsh to prevent the above
  nightmare from recurring.