Rewrite `git-create-pull-request` in Ruby
As part of this rewrite, the script got some hot new features.

The biggest new feature is that it reflows commit messages (in a
Markdown-aware way) to not have linebreaks, since GitHub keeps those
hard line breaks and it looks bad.

It has a Hired-specific feature where it inserts commits in the correct
section of the `PULL_REQUEST_TEMPLATE` rather than sticking them at the
bottom of the file.

Fixes #73.