Load plugins later in the Vim boot process
I ran into a problem with my rainbow parentheses plugin, where I was
conditionally setting a variable:

    if expand("%:e") ==# 'clj'
      let g:rainbow_active = 1
    endif

However, I was setting `g:rainbow_active` _after_ the plugin loaded, and
so it never read it.

To fix this problem forever (and possibly for other plugins too, in the
future), load plugins as late as possible.

There are a few things that must happen _after_ the plugins:

* `colorscheme`
* `filetype plugin syntax on`

But everything else is lazy-loaded and so loading the plugins later
won't affect it.