Don't check for Ale
The `ale#Lint` function is defined after a file loads, I think, so even
if Ale is installed, the function isn't defined at the time the `if` is
evaluated. (I tried moving the `if` farther down my vimrc but it didn't
change anything.)

The whole point of the `if` was to prevent Vim from crashing when
running on a new computer. However, `install.sh` runs `vim +PlugInstall`
just fine without Ale, so it can bootstrap itself just fine.

All of this points towards Just Removing The If.