Sort git branches by commit date
The author date (`authordate` key) on a commit is preserved on rebase /
cherry-pick etc, but the commit date (`committerdate` key) is changed.

I want the commit date.

(I prefix it with `-` so the most recent branches are at the top.)

Thank you to this blog post for telling me about branch sorting:
https://blog.github.com/2018-09-10-highlights-from-git-2-19/