t: Use regular FZF most of the time
When I call `t` in tmux (<Prefix> + t), I'm usually in another program,
like Vim, and so popping open a new tmux pane with fzf-tmux is a good
idea. But when I call `t` directly in Zsh, it should just open in the
current pane, instead of further splitting the current, available pane.